+++ src/lando.ts
@@ -1,9 +1,17 @@
-import { window, commands } from 'vscode';
-import { outputChannel, getAppNameFromAppConfig, getLandoFile, getCurrentAppName } from './extension';
+import { window, commands, workspace, Uri } from 'vscode';
+import {
+  outputChannel,
+  getAppNameFromAppConfig,
+  getLandoFile,
+  getCurrentAppName,
+  getWorkspaceFolderPath,
+  getWorkspaceFolderNameFromPath
+} from './extension';
 import { exec, execSync } from 'child_process';
 import * as json from 'jsonc-parser';
 import * as stripAnsi from 'strip-ansi';
 import { setButtonTo, showOutput } from './commands';
+import { fstat, unlink } from 'fs';
 
 export function init(): void {
   var terminal = window.createTerminal('Lando Init');
@@ -175,13 +183,100 @@ export function version(): string {
   return stdout;
 }
 
+export function dbExport(dir: string, host?: string, filePath?: string) {
+  let command = 'lando db-export';
+  if (host) command += ' -h ' + host;
+  if (filePath) command += ' "' + filePath + '"';
+
+  const child = exec(command, { cwd: dir });
+  child.stdout.on('data', data => {
+    if (data.includes('Could not find app in this dir or a reasonable amount of directories above it!')) {
+      window.showWarningMessage('Please initiate a lando project: ' + data);
+      setButtonTo('init');
+    }
+    if (data.includes('Failed')) {
+      window.showWarningMessage('Failed to export database from ' + host);
+    }
+    if (data.includes('Success')) {
+      window.showInformationMessage('Successfully exported database from ' + host);
+    }
+    outputChannel.append(`${stripAnsi.default(data)}`);
+  });
+  child.stderr.on('data', data => {
+    outputChannel.append(`${stripAnsi.default(data)}`);
+  });
+  child.on('exit', (code, signal) => {
+    outputChannel.appendLine('-----------------------');
+    outputChannel.appendLine('child process exited with ' + `code ${code} and signal ${signal}`);
+    outputChannel.appendLine('-----------------------');
+  });
+}
+
+export function dbExportOut(dir: string, host?: string) {
+  let command = 'lando db-export --stdout';
+  if (host) command += ' -h ' + host;
+  var stdout = execSync(command, { encoding: 'utf8' });
+  return stdout;
+}
+
+export function dbImport(dir: string, host?: string, noWipe?: boolean, filePath?: string, isTmp?: boolean) {
+  let command = 'lando db-import';
+  if (host) command += ' -h ' + host;
+  if (noWipe) command += ' --no-wipe';
+  if (filePath) command += ' "' + filePath + '"';
+
+  const child = exec(command, { cwd: dir });
+  child.stdout.on('data', data => {
+    if (data.includes('Could not find app in this dir or a reasonable amount of directories above it!')) {
+      window.showWarningMessage('Please initiate a lando project: ' + data);
+      setButtonTo('init');
+    }
+    if (data.includes('not found')) {
+      window.showWarningMessage('Failed to import database into ' + host);
+    }
+    if (data.includes('Preparing to import')) {
+      window.showInformationMessage('Importing database into ' + host);
+    }
+    if (data.includes('Import complete')) {
+      window.showInformationMessage('Successfully imported database to ' + host);
+    }
+    outputChannel.append(`${stripAnsi.default(data)}`);
+  });
+  child.stderr.on('data', data => {
+    outputChannel.append(`${stripAnsi.default(data)}`);
+  });
+  child.on('exit', (code, signal) => {
+    outputChannel.appendLine('-----------------------');
+    outputChannel.appendLine('child process exited with ' + `code ${code} and signal ${signal}`);
+    outputChannel.appendLine('-----------------------');
+    if (isTmp) {
+      unlink(getWorkspaceFolderPath() + '/' + filePath, err => {
+        if (err) {
+          window.showErrorMessage('Failed to remove tmp import file');
+          outputChannel.appendLine('-----------------------');
+          outputChannel.appendLine('Failed to remove tmp import file: ' + err);
+          outputChannel.appendLine('-----------------------');
+          return;
+        }
+      });
+    }
+  });
+}
+
+export function sshService(offset: number, provider: any): void {
+  var treeItem = provider.getTreeItem(offset);
+  var terminal = window.createTerminal('Lando ' + treeItem.label);
+  terminal.show();
+  terminal.sendText('lando ssh -s ' + treeItem.label);
+}
+
 export function reformatInfo(info: string): string {
   var newInfo: { [k: string]: any } = {};
   var infoJson = json.parse(info);
   infoJson.forEach((element: any) => {
     var service = element.service;
     delete element.service;
-    newInfo[service] = element;
+    newInfo['service_' + service] = element;
   });
   return JSON.stringify(newInfo);
 }